Add files via upload
To run this project, open the project in the VS Code and open a new terminal . Write "npm i" in the terminal and press enter ,it will create you node modules folder. Now to see the website write "npm run start" in the terminal and you are good to go.